Dougbot.Twitch/requirements.txt
discord.py==2.3.2
twitchAPI==4.2.0
motor==3.3.2
pytz==2024.1